Add test for subsequent sessions
Most of the existing tests only interacted with a fresh instance of
ammo. With the addition of this test, there is reasonable assurance that
it will save and load resources correctly, and will generally behave the
same regardless of whether it's the first launch or hundredth launch.

+++ test/common.py
@@ -68,9 +68,12 @@ class AmmoController:
                         # directory wasn't empty, ignore this
                         pass
 
-        remove_empty_dirs(self.game_dir)
-        shutil.rmtree(self.game_dir)
-        shutil.rmtree(self.ammo_dir)
+        if os.path.exists(self.game_dir):
+            remove_empty_dirs(self.game_dir)
+        if os.path.exists(self.game_dir):
+            shutil.rmtree(self.game_dir)
+        if os.path.exists(self.ammo_dir):
+            shutil.rmtree(self.ammo_dir)
 
 
 def mod_extracts_files(mod_name, files):
